fixed rendering of tables, code. added copy, fixed error for settings sync, added additional formatting and styles to chat
