+++ apps/sim/app/chat/[subdomain]/components/header-links/links.tsx
@@ -12,16 +12,17 @@ export default function HeaderLinks({ stars }: HeaderLinksProps) {
     <div className="flex items-center">
       <motion.a
         href="https://github.com/simstudioai/sim"
-        className="flex items-center gap-1.5 text-foreground/80 hover:text-foreground/100 p-1 rounded-md transition-colors duration-200"
+        className="flex items-center gap-1 text-foreground/70 hover:text-foreground transition-colors duration-200 rounded-md px-1.5 py-1 hover:bg-foreground/5"
         aria-label="GitHub"
         target="_blank"
         rel="noopener noreferrer"
         initial={{ opacity: 0, y: -5 }}
         animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.2, ease: 'easeOut', delay: 0.1 }}
+        transition={{ duration: 0.3, ease: 'easeOut' }}
+        whileHover={{ scale: 1.02 }}
       >
-        <GithubIcon className="w-[24px] h-[24px]" />
-        <span className="text-sm font-medium hidden sm:inline">{stars}</span>
+        <GithubIcon className="w-[18px] h-[18px]" />
+        <span className="text-xs font-medium hidden sm:inline-block">{stars}</span>
       </motion.a>
     </div>
   )
